add bigdl 4
